pyproject.toml
[project]
name = "software-engineer-project-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
]
